Move `Logger` classes to `Log` subfolder
