+++ srcStatic/Log/Logger.h
