+++ srcStatic/Log/LoggerDebug.cpp
