+++ srcStatic/Log/LoggerDebug.h
