+++ srcStatic/Log/LoggerDistributor.cpp
