+++ srcStatic/Log/LoggerFile.cpp
@@ -1,7 +1,7 @@
 #include "LoggerFile.h"
-#include "Log.h"
-#include "FileSystemHelper.h"
-#include "StringConversion.h"
+#include "../Log.h"
+#include "../FileSystemHelper.h"
+#include "../StringConversion.h"
 #include <cstddef>
