+++ srcStatic/op2extStatic.vcxproj.filters
@@ -7,10 +7,10 @@
     <ClCompile Include="ModuleLoader.cpp" />
     <ClCompile Include="IniFile.cpp" />
     <ClCompile Include="Log.cpp" />
-    <ClCompile Include="LoggerDebug.cpp" />
-    <ClCompile Include="LoggerDistributor.cpp" />
-    <ClCompile Include="LoggerFile.cpp" />
-    <ClCompile Include="LoggerMessageBox.cpp" />
+    <ClCompile Include="Log/LoggerDebug.cpp" />
+    <ClCompile Include="Log/LoggerDistributor.cpp" />
+    <ClCompile Include="Log/LoggerFile.cpp" />
+    <ClCompile Include="Log/LoggerMessageBox.cpp" />
     <ClCompile Include="OP2Memory.cpp" />
     <ClCompile Include="ResourceSearchPath.cpp" />
     <ClCompile Include="StringConversion.cpp" />
@@ -41,11 +41,11 @@
     <ClInclude Include="IniFile.h" />
     <ClInclude Include="LocalResource.h" />
     <ClInclude Include="Log.h" />
-    <ClInclude Include="Logger.h" />
-    <ClInclude Include="LoggerDebug.h" />
-    <ClInclude Include="LoggerDistributor.h" />
-    <ClInclude Include="LoggerFile.h" />
-    <ClInclude Include="LoggerMessageBox.h" />
+    <ClInclude Include="Log/Logger.h" />
+    <ClInclude Include="Log/LoggerDebug.h" />
+    <ClInclude Include="Log/LoggerDistributor.h" />
+    <ClInclude Include="Log/LoggerFile.h" />
+    <ClInclude Include="Log/LoggerMessageBox.h" />
     <ClInclude Include="OP2Memory.h" />
     <ClInclude Include="ResourceSearchPath.h" />
     <ClInclude Include="StringConversion.h" />
