+++ test/LogMessageTest.h
@@ -1,4 +1,4 @@
-#include "Logger.h"
+#include "Log/Logger.h"
 #include <gtest/gtest.h>
 #include <gmock/gmock.h>
